Gave all components their own homes
Moved all components in the old `src/xsd.rs` to submodules.
Exposed all components in mod.rs for a nice clean import path.
Ran cargo clippy/fmt.
