+++ src/xsd/attribute.rs
@@ -1,0 +1,6 @@
+#[derive(Debug)]
+pub struct Attribute {
+    pub name: String,
+    pub r#type: String,
+    pub r#use: String,
+}
